fix: switch to magic file validator by default
